requirements.txt
Django>=4
djangorestframework
djangorestframework-simplejwt
django-oauth-toolkit